% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{pivot_help_UI}
\alias{pivot_help_UI}
\title{PIVOT help modules, UI}
\usage{
pivot_help_UI(
  id,
  title,
  label = NULL,
  icn = "question-circle",
  type = "button",
  tooltip = T,
  style = NULL
)
}
\description{
PIVOT help modules, UI
}
